closes #15 Implement Shared UI Component Library (Tailwind/Glassmorphism)
